part4

+++ src/CPG.Application/UseCases/Providers/Exceptions/ProviderNotFoundException.cs
@@ -2,8 +2,8 @@
 
 namespace CPG.Application.UseCases.Providers.Exceptions;
 
-public class ProviderNotFoundException(int providerId) : ApplicationException($"Provider with ID {providerId} has not been found.")
+public class ProviderNotFoundException(long providerId) : ApplicationException($"Provider with ID {providerId} has not been found.")
 {
     public override string Code => "provider_not_found";
-    public int ProviderId { get; } = providerId;
+    public long ProviderId { get; } = providerId;
 }
